requirements.txt
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.16.3
opencv-python==4.8.0.74
numpy<2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.8.3
google-generativeai==0.8.3
